Extract sync payload codec
