+++ js/sync-delta-row-codec.js
@@ -1,7 +1,7 @@
 // @ts-check
 // sync-delta-row-codec.js - Shared itemRow payload decoding for delta merge paths.
 
-import { _base64ToBytes, _gunzipToStringCapped } from './sync-payload.js';
+import { _base64ToBytes, _gunzipToStringCapped } from './sync-payload-codec.js';
 
 export async function decodeRowPayload(row) {
   let json = row.payload;
