+++ scripts/architecture-cycle-baseline.json
@@ -1,7 +1,7 @@
 {
   "schemaVersion": 1,
-  "maxCyclicModules": 63,
-  "maxLargestCyclicComponent": 54,
+  "maxCyclicModules": 49,
+  "maxLargestCyclicComponent": 40,
   "allowedCyclicModules": [
     "js/api-models.js",
     "js/api-openai-compatible.js",
@@ -29,19 +29,6 @@
     "js/sync-chat-apply.js",
     "js/sync-configure.js",
     "js/sync-cutover.js",
-    "js/sync-delta-array-merge.js",
-    "js/sync-delta-array-planner.js",
-    "js/sync-delta-map-merge.js",
-    "js/sync-delta-map-planner.js",
-    "js/sync-delta-merge-shapes.js",
-    "js/sync-delta-merge.js",
-    "js/sync-delta-planners.js",
-    "js/sync-delta-row-codec.js",
-    "js/sync-delta-scalar-merge.js",
-    "js/sync-delta-scalar-planner.js",
-    "js/sync-delta.js",
-    "js/sync-diagnose-actions.js",
-    "js/sync-diagnose-cutover-actions.js",
     "js/sync-diagnose-ui.js",
     "js/sync-diagnostics-snapshot.js",
     "js/sync-diagnostics.js",
@@ -53,7 +40,6 @@
     "js/sync-pull-active-refresh.js",
     "js/sync-pull-merge.js",
     "js/sync-pull.js",
-    "js/sync-push-deltas.js",
     "js/sync-push.js",
     "js/sync-reconcile.js",
     "js/sync-save-hooks.js",
