+++ service-worker.js
@@ -409,6 +409,7 @@ const APP_SHELL = [
   '/js/sync-cutover.js',
   '/js/sync-ui.js',
   '/js/sync-payload-collectors.js',
+  '/js/sync-payload-codec.js',
   '/js/sync-payload.js',
   '/js/sync-relay-health.js',
   '/js/sync-state.js',
